+++ tests/Client/ClientErrorsTest.php
@@ -118,7 +118,7 @@ class ClientErrorsTest extends TestCase
                 ->willReturn($response);
         }
 
-        return new MtgJsonClient($httpClientMock, $httpFactory, $httpFactory, $httpFactory, new NullLogger());
+        return new MtgJsonClient($httpClientMock, $httpFactory, $httpFactory, new NullLogger());
     }
 
     /**